+++ packages/schema-blocks/src/core/validation/BlockValidation.ts
@@ -1,4 +1,6 @@
 export enum BlockValidation {
+	/** This schema block is defined to recommend a particular inner block, but that block doesn't exist. */
+	MissingRecommendedBlock = -3,
 	/** This block was skipped during validation, on purpose. If you ever see this value, that's a bug.  */
 	Skipped = -2,
 	/** This block doesn't have any code that validates it, so we simply don't know if it's valid or not.  */
@@ -10,7 +12,7 @@ export enum BlockValidation {
 	/** This block has required attributes, but these attributes are missing or empty. */
 	MissingAttribute = 2,
 	/** This schema block is defined to require a particular inner block, but that block doesn't exist. */
-	MissingBlock = 3,
+	MissingRequiredBlock = 3,
 	/** There may be only one of this type of block, but we found more than one. */
 	TooMany = 4,
 }