+++ packages/schema-blocks/src/functions/presenters/BlockSuggestionsPresenter.tsx
@@ -40,7 +40,7 @@ function BlockSuggestion( { blockTitle, blockName, blockClientId }: BlockSuggest
 	};
 
 	return (
-		<li className="yoast-block-suggestion">
+		<li className="yoast-block-suggestion" key={ blockTitle }>
 			{ blockTitle }
 			<button className="yoast-block-suggestion-button" onClick={ addBlockClick }> Add </button>
 		</li>
@@ -56,7 +56,7 @@ function BlockSuggestion( { blockTitle, blockName, blockClientId }: BlockSuggest
  */
 function BlockSuggestionAdded( { blockTitle }: BlockSuggestionAddedDto ): ReactElement {
 	return (
-		<li className="yoast-block-suggestion yoast-block-suggestion--added">
+		<li className="yoast-block-suggestion yoast-block-suggestion--added" key={ "BlockSuggestionAdded" + blockTitle }>
 			{ blockTitle }
 			<span className="yoast-block-suggestion-checkmark"> OK </span>
 		</li>
@@ -83,7 +83,7 @@ export default function BlockSuggestionsPresenter( { title, block, suggestions }
 	const presentBlockNames = findPresentBlocks.map( presentBlock => presentBlock.name );
 
 	return (
-		<PanelBody key={ block.clientId }>
+		<PanelBody key={ title + block.clientId }>
 			<div className="yoast-block-sidebar-title">{ title }</div>
 			<ul className="yoast-block-suggestions">
 				{