+++ packages/schema-blocks/tests/functions/presenters/SidebarWarningPresenter.test.ts
@@ -46,9 +46,26 @@ describe( "The SidebarWarningPresenter ", () => {
 			} ] );
 		} );
 
+		it( "creates a footer message when some required blocks have missing attributes.", () => {
+			const testcase = new BlockValidationResult( "1", "mijnblock", BlockValidation.Invalid, BlockPresence.Unknown );
+			testcase.issues.push(
+				new BlockValidationResult( null, "missingblockattribute", BlockValidation.MissingAttribute, BlockPresence.Unknown )
+			);
+
+			const result = createAnalysisMessages( testcase );
+
+			expect( result.length ).toEqual( 1 );
+			expect( result[ 0 ] ).toEqual(
+				{
+					text: "Not all required blocks have been completed! No mijnblock schema will be generated for your page.",
+					color: "red",
+				},
+			);
+		} );
+
 		it( "creates warning messages for missing required blocks, with a footer message.", () => {
 			const testcase = new BlockValidationResult( "1", "mijnblock", BlockValidation.Invalid, BlockPresence.Required );
-			testcase.issues.push( BlockValidationResult.MissingBlock( "missingblock", BlockPresence.Required ) );
+			testcase.issues.push( BlockValidationResult.MissingRequiredBlock( "missingblock" ) );
 
 			const result = createAnalysisMessages( testcase );
 
@@ -68,10 +85,10 @@ describe( "The SidebarWarningPresenter ", () => {
 			"the conclusion should still be green.", () => {
 			const testcase = new BlockValidationResult( "1", "mijnblock", BlockValidation.Invalid, BlockPresence.Recommended );
 			testcase.issues.push(
-				BlockValidationResult.MissingBlock( "missing recommended block", BlockPresence.Recommended ),
+				BlockValidationResult.MissingRecommendedBlock( "missing recommended block" ),
 			);
 			testcase.issues.push(
-				BlockValidationResult.MissingBlock( "missing recommended block 2", BlockPresence.Recommended ),
+				BlockValidationResult.MissingRecommendedBlock( "missing recommended block 2" ),
 			);
 
 			const result = createAnalysisMessages( testcase );
@@ -115,13 +132,13 @@ describe( "The SidebarWarningPresenter ", () => {
 
 			expect( result ).toEqual( [ {
 				text: "Good job! All required blocks have been completed.",
-				color: "green"
+				color: "green",
 			} ] );
 		} );
 
 		it( "creates a warning for a required block with validation problems.", () => {
 			const testcase = new BlockValidationResult( "1", "myBlock", BlockValidation.Invalid, BlockPresence.Required );
-			testcase.issues.push( BlockValidationResult.MissingBlock( "innerblock1", BlockPresence.Required ) );
+			testcase.issues.push( BlockValidationResult.MissingRequiredBlock( "innerblock1" ) );
 			validations[ "1" ] = testcase;
 
 			const result = getWarnings( "1" );